Fixed some atmosphere bugs
Mie is now properly affecting the background

Prevented luminance from passing a NAN value to the scene

+++ Raytracer/Raytracer/camera.hpp
@@ -49,7 +49,7 @@ public:
     /// <returns>Color correction scale factor</returns>
     float AdaptiveLogTR(float bias, float adaptaionLuminance, float maxLuminanceDisplay, float worldLuminance);
 
-    float LogAverageLuminance();
+    double LogAverageLuminance();
 
     Camera(float focalLen, float fov, glm::vec3 pos);
     ~Camera();
